Add filter actions (auto mark read) at global level fix https://github.com/FreshRSS/FreshRSS/issues/2788
