+++ app/Models/Tag.php
@@ -2,13 +2,10 @@
 declare(strict_types=1);
 
 class FreshRSS_Tag extends Minz_Model {
+	use FreshRSS_AttributesTrait;
 
 	private int $id = 0;
 	private string $name;
-	/**
-	 * @var array<string,mixed>
-	 */
-	private array $attributes = [];
 	private int $nbEntries = -1;
 	private int $nbUnread = -1;
 
@@ -33,34 +30,6 @@ class FreshRSS_Tag extends Minz_Model {
 
 	public function _name(string $value): void {
 		$this->name = trim($value);
-	}
-
-	/**
-	 * @phpstan-return ($key is non-empty-string ? mixed : array<string,mixed>)
-	 * @return array<string,mixed>|mixed|null
-	 */
-	public function attributes(string $key = '') {
-		if ($key === '') {
-			return $this->attributes;
-		} else {
-			return $this->attributes[$key] ?? null;
-		}
-	}
-
-	/** @param string|array<mixed>|bool|int|null $value Value, not HTML-encoded */
-	public function _attributes(string $key, $value = null): void {
-		if ($key == '') {
-			if (is_string($value)) {
-				$value = json_decode($value, true);
-			}
-			if (is_array($value)) {
-				$this->attributes = $value;
-			}
-		} elseif ($value === null) {
-			unset($this->attributes[$key]);
-		} else {
-			$this->attributes[$key] = $value;
-		}
 	}
 
 	public function nbEntries(): int {
