+++ app/views/configure/reading.phtml
@@ -346,6 +346,21 @@
 		</fieldset>
 
 		<fieldset>
+			<legend><?= _t('sub.feed.filteractions') ?></legend>
+			<div class="form-group">
+				<label class="group-name" for="filteractions_read"><?= _t('conf.reading.read.when') ?></label>
+				<div class="group-controls">
+					<textarea name="filteractions_read" id="filteractions_read" class="w100"><?php
+						foreach (FreshRSS_Context::$user_conf->filtersAction('read') as $filterRead) {
+							echo $filterRead->getRawInput(), PHP_EOL;
+						}
+					?></textarea>
+					<p class="help"><?= _i('help') ?> <?= _t('sub.feed.filteractions.help') ?></p>
+				</div>
+			</div>
+		</fieldset>
+
+		<fieldset>
 			<legend><?= _t('conf.reading.headline.misc') ?></legend>
 			<div class="form-group">
 				<div class="group-controls">
