+++ config-user.default.php
@@ -64,6 +64,7 @@ return array (
 		'focus' => false,
 		'site' => true,
 	),
+	'filters' => [],
 	'theme' => 'Origine',
 	'darkMode' => 'no',
 	'content_width' => 'thin',
